3dot and dropupmenu modularized
